# Improve foldr performance ~4x on large lists