@milesfrain 2020-12-20
This slower reverse pattern is duplicated in a few other places (just search for `foldl (flip`, for example:
https://github.com/purescript/purescript-lists/blob/62900a56f6864638c952575dfd211a1cc55be7da/src/Data/List/Types.purs#L135
https://github.com/purescript/purescript-lists/blob/62900a56f6864638c952575dfd211a1cc55be7da/src/Data/List/Types.purs#L141
https://github.com/purescript/purescript-lists/blob/62900a56f6864638c952575dfd211a1cc55be7da/src/Data/List/Types.purs#L145
https://github.com/purescript/purescript-lists/blob/62900a56f6864638c952575dfd211a1cc55be7da/src/Data/List/Types.purs#L153
https://github.com/purescript/purescript-lists/blob/62900a56f6864638c952575dfd211a1cc55be7da/src/Data/List/Types.purs#L260

So thinking these should all be replaced with a common `rev` helper function in this file (along with some comments about why it's written longform).

I'm also wondering if we should do more benchmarking in the browser (blocked by https://github.com/purescript/purescript-minibench/issues/16) with bundling optimizations and see if there's still a performance difference. Ideally, I'd like to keep the original concise code.